requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
astor==0.8.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
grpcio==1.54.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.5.0
joblib==1.2.0
json-tricks==3.16.1
lightning-utilities==0.5.0
Markdown==3.4.3
MarkupSafe==2.1.2
multidict==6.0.4
nni==2.10
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
prettytable==3.6.0
protobuf==3.20.1
psutil==5.9.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytorch-lightning==1.8.6
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.22.0
rsa==4.9
schema==0.7.5
scikit-learn==1.2.0
scipy==1.10.0
simplejson==3.18.1
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
timm==0.6.12
toml==0.10.2
tqdm==4.64.1
typeguard==2.13.3
types-toml==0.10.8.1
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
